requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
razorpay
python-dotenv
Flask-Mail
email-validator
requests